standalone/experiments/multi-agent/multi-agent-system/requirements.txt
strands-agents>=0.1.0
strands-agents-tools[a2a_client]>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
httpx>=0.25.0
a2a>=0.1.0